pyproject.toml
[project]
name = "bdbot"
version = "4.8.1"
description = "A discord bot that post comics strip."
readme = "README.md"
authors = [{name = "Arik", email = "BBArik@protonmail.com"}]
license = {text = "GPL 3.0"}
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.5.1",
    "requests>=2.32.3",
    "requests-htmlc>=0.0.8",
    "rss-parser>=2.1.0",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.1",
    "randomtimestamp @ git+https://github.com/ByteBaker/randomtimestamp.git",
    "topggpy>=1.4.0",
    "inquirerpy>=0.3.4",
    "dotenv",
    "tortoise-orm",
    "tortoise-cli"
]

[project.optional-dependencies]
dev = ["pytest>=8.3.4", "flake8>=7.1.2", "isort>=6.0.1", "black>=25.1.0", "pre-commit>=4.1.0", "Flake8-pyproject>=1.2.3"]


[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.scripts]
bdbot = "bdbot.__main__:main"
bdbot_manager = 'bdbot.bdbot_manager:main'

[tool.setuptools]
packages = ["bdbot"]

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", ".venv", "venv"]
